moved logic from main to game manager
